pyproject.toml
[project]
name = "ai-for-lawyers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "google-genai>=1.47.0",
    "matplotlib>=3.9.4",
    "ollama>=0.6.1",
    "openai>=2.29.0",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "scikit-learn>=1.6.1",
    "streamlit>=1.50.0",
]
